+++ src/components/properties/TypographySection.vue
@@ -4,6 +4,7 @@ import { ToggleGroupItem, ToggleGroupRoot } from 'reka-ui'
 import { TypographyControlsRoot, useI18n } from '@open-pencil/vue'
 
 import FontPicker from '@/components/FontPicker.vue'
+import FontSettingsPopover from '@/components/FontSettingsPopover.vue'
 import ScrubInput from '@/components/ScrubInput.vue'
 import VariableScrubInput from '@/components/properties/VariableScrubInput.vue'
 import AppSelect from '@/components/ui/AppSelect.vue'
@@ -27,6 +28,7 @@ const fontLoader = { load: loadFont }
           :model-value="ctx.node.value.fontFamily"
           @select="ctx.actions.setFamily"
         />
+        <FontSettingsPopover />
         <Tip
           v-if="ctx.hasMissingFonts.value"
           :label="